pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["SevaErsh <erschidvs@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.2"
Jinja2 = "^3.1.2"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.3"
django-filter = "^22.1"
rollbar = "^0.16.3"

[tool.poetry.dev-dependencies]
heroku = "^0.1.4"
pip = "^22.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
